Revamp novehicle claimflag (#130)
* Revamp novehicle claimflag

1: Fixes an exploit where if two players who do not have permission in a claim ride a boat into a claim with novehicle, they can duplicate the boat. Both of them would have the boat in their inventory.
2. Since griefprevention already disables placement of vehicles in a claim for players without permission, I have removed that part of the code
3. Instead of breaking the vehicle on teleport into a novehicle claim, this will just remove the player from the vehicle, and then teleport just the player.
4. The claim owner used to not have permission to ride a boat in their own novehicle claim. This grants them that ability. Same for mounting.
5. If the player driving the boat has permission in the claim, allow them to continue riding it even if the second passenger doesn't.
6. Prevents all pushing of vehicles inside of novehicle claims.

This has been tested on my own server for several minutes without any issues.

I think a possible disadvantage for these changes is just that the way that the novehicle flag works has been completely revamped since the original 5.9.3. The one part of it that was working (preventing vehicle placement) was removed and a whole bunch of new stuff was added. Perhaps, it might make more sense to revert the novehicle flag to what it used to be in the original 5.9.3 and make a new claimflag for all of these changes.

* Small changes

1. When a player teleports with a vehicle into a novehicle claimflag, send them a message
2. Fix NPE when a vehicle with no passengers moves

* Fix one last exception

* Requested Changes

1. Removed redundant null check
2. Made eventName boolean

* Allow claim members to push vehicles

* use passengers instead of vehicle.getPassengers();